fix(gcc13): explicit cast on ARRAYLEN operand in dshot safeCount clamp
pwm_output.c:420 - MIN(motorCount, ARRAYLEN(...)) mixes uint8_t and
size_t operands; the GCC statement-expression MIN macro promotes both
to a common type before narrowing back to uint8_t on assignment. Not
currently flagged (EF does not enable -Wconversion/-Wsign-conversion),
but an explicit (uint8_t) cast on the ARRAYLEN operand removes the
implicit narrowing entirely, defensively, at zero behavior change.

Confirmed via second-opinion review (opencode deepseek-v4-flash-free)
and comparison against Betaflight's dshot_command.c, which has never
needed this clamp at all (motorCount is unclamped there) — this
remains strictly defensive hardening, not a fix for a reachable bug.

+++ src/main/drivers/pwm_output.c
@@ -417,7 +417,7 @@ void pwmWriteDshotCommand(uint8_t index, uint8_t motorCount, uint8_t command, bo
         }
         delayMicroseconds(delayAfterCommandUs);
     } else {
-        const uint8_t safeCount = MIN(motorCount, ARRAYLEN(dshotCommandControl.command));
+        const uint8_t safeCount = MIN(motorCount, (uint8_t)ARRAYLEN(dshotCommandControl.command));
         dshotCommandControl.repeats = repeats;
         dshotCommandControl.nextCommandAtUs = timeNowUs + DSHOT_INITIAL_DELAY_US;
         dshotCommandControl.delayAfterCommandUs = delayAfterCommandUs;
